+++ Q3/PacletInfo.m
@@ -1,18 +1,16 @@
 (* Paclet Info File *)
 
-(* created 2021/01/11*)
+(* created 2021/01/15*)
 
 Paclet[
     Name -> "Q3",
-    Version -> "1.4.0",
-    MathematicaVersion -> "6+",
+    Version -> "1.4.1",
+    MathematicaVersion -> "12+",
     Description -> "Mathematica application to help study quantum information processing, quantum many-body systems, and quantum spin systems.",
     Creator -> "Mahn-Soo Choi (Korea University)",
     Extensions -> 
         {
-            {"Documentation", Resources -> 
-                {"Guides/Q3"}
-            , Language -> "English"}
+          { "Documentation", Language -> "English" }
         }
 ]